added type requirement for autodiff
produces a more graceful error when forwarddiff_derivative is called on
something other than a Float64

+++ src/activations.jl
@@ -25,6 +25,6 @@ derivs = Dict{Function, Function}([logis     => logisd,
 # will only produce derivatives with Float64 methods
 function autodiff(activ::Function)
     f(x) = activ(x[1])
-    forwarddiff_derivative(x) = forwarddiff_gradient(f,Float64)([x])[1]
+    forwarddiff_derivative(x::Float64) = forwarddiff_gradient(f,Float64)([x])[1]
     return forwarddiff_derivative
 end
